requirements.txt
google-generativeai
requests
gradio
fpdf2
python-dotenv